Remove HttpInterceptorContext class from Styx server.
